requirements.txt
python-dotenv==0.21.1
pyspark==3.4.1
boto3==1.28.79
urllib3==1.26.9
pandas==1.3.5
```